+++ mocks/Cargo.toml
@@ -15,7 +15,7 @@ resolver = "2"
 prost = { workspace = true }
 serde_json = { workspace = true }
 tokio = { version = "1", features = ["full"] }
-tower-abci = "0.10"
+tower-abci = "0.11"
 tower = { version = "0.4", features = ["full"] }
 tempfile = { workspace = true }
 
@@ -25,9 +25,9 @@ sov-ibc-transfer = { path = "../modules/sov-ibc-transfer" }
 
 # ibc dependencies
 ibc = { workspace = true }
-ibc-proto = { version = "0.35", default-features = false, features = ["borsh"] }
-basecoin-app = { git = "https://github.com/informalsystems/basecoin-rs.git", rev = "5f39639" }
-basecoin-store = { git = "https://github.com/informalsystems/basecoin-rs.git", rev = "5f39639" }
+ibc-proto = { version = "0.37.1", default-features = false, features = ["borsh"] }
+basecoin-app = { git = "https://github.com/informalsystems/basecoin-rs.git", rev = "823f846" }
+basecoin-store = { git = "https://github.com/informalsystems/basecoin-rs.git", rev = "823f846" }
 tendermint = { workspace = true }
 tendermint-testgen = { workspace = true }
